fix autoloader precedence
Since Composer uses $GLOBALS under the hood to determins
which classes have been loaded, prefer the project-wide autoloader
over the plugin/local one.

This assumes that users tracking composer dependencies in their project
repo will pull in GearLab's dependencies explicitly. This isn't ideal,
but this is a very advanced use-case so assuming this level of care is
acceptable.